Create app component
